Move ShotList to its own file
