+++ src/gui/src/main.cpp
@@ -28,6 +28,7 @@
 #include <core/Version.h>
 #include <getopt.h>
 
+#include "ShotList.h"
 #include "SplashScreen.h"
 #include "HydrogenApp.h"
 #include "MainForm.h"
@@ -191,248 +192,6 @@ public:
 };
 
 
-/// Shot List
-///
-/// Utility for grabbing screenshots of widgets. Not a script, just a shot list.
-///
-/// Entries are a space-separated list of the form:
-///
-///  grab <WidgetName|WidgetClass>   -- grab widget named "WidgetName" or of any class inheriting from "WidgetClass"
-///  slot <widget> <slot> [<arg>...] -- invoke a slot method on widget
-///  fin                             -- quit Hydrogen
-///  dump                            -- dump object tree(s)
-///  # text                          -- commentary
-///
-class ShotList : public QObject {
-
-	static QWidget *findWidgetInheriting( QObject *pObject, QString &sName ) {
-		if ( pObject->inherits( sName.toLocal8Bit().data() ) ) {
-			return dynamic_cast< QWidget *>( pObject );
-		}
-		for ( QObject *pC : pObject->children() ) {
-			QWidget *pW = findWidgetInheriting( pC, sName );
-			if ( pW ) {
-				return pW;
-			}
-		}
-		return nullptr;
-	}
-
-	static QWidget *findWidget( QString &sName ) {
-		for ( QWidget * pTop : QApplication::topLevelWidgets() ) {
-			QWidget *pWidget = pTop->findChild< QWidget *>( sName );
-			if ( !pWidget && pTop->objectName() == sName ) {
-				pWidget = dynamic_cast< QWidget *>( pTop );
-			}
-			if ( !pWidget ) {
-				pWidget = findWidgetInheriting( pTop, sName );
-			}
-			if ( pWidget ) {
-				return pWidget;
-			}
-		}
-		return nullptr;
-	}
-
-	/// Buffer for construction of Q_ARGs
-	class Arg {
-		QString m_sArg;
-		union {
-			int m_n;
-			QWidget *m_pWidget;
-		};
-	public:
-		Arg( QString &sArg ) : m_sArg( sArg ) {}
-
-		operator QGenericArgument() {
-			if (( m_sArg == "true" )) {
-				return Q_ARG( bool, true );
-			} else if (( m_sArg == "false" )) {
-				return Q_ARG( bool, false );
-			}
-			bool bIsInt = false;
-			m_n = m_sArg.toInt( &bIsInt );
-			if ( bIsInt ) {
-				return Q_ARG( int, m_n );
-			} else if (( m_pWidget = findWidget( m_sArg ) )) {
-				return Q_ARG( QWidget *, m_pWidget );
-			} else {
-				// Last resort, treat as a QString
-				return Q_ARG( QString, m_sArg );
-			}
-		}
-	};
-
-
-private:
-	QStringList m_shots;
-	int m_nNextShot;
-
-	void shoot( QString s ) {
-		___INFOLOG( QString( "Taking shot: %1" ).arg( s.trimmed() ) );
-		QStringList words = s.trimmed().split( QRegExp( "\\s+" ) );
-		if ( s.size() == 0 ) {
-			return;
-		}
-		QString sCmd = words[ 0 ];
-
-		if ( sCmd == "#" || sCmd == "" ) {
-			// Empty line or "#" to start a comment
-		} else if ( sCmd.compare( "fin", Qt::CaseInsensitive) == 0 ) {
-			// Finish the shot list and quit Hydrogen
-			QTimer::singleShot( 1, QApplication::instance(), &QApplication::closeAllWindows );
-	    } else if ( sCmd.compare( "dump", Qt::CaseInsensitive) == 0 ) {
-			// Dump object tree for debugging
-			for ( QWidget *pTop : QApplication::topLevelWidgets() ) {
-				pTop->dumpObjectTree();
-			}
-		} else if ( sCmd.compare( "grab", Qt::CaseInsensitive ) == 0 ) {
-
-			if ( words.size() < 2 ) {
-				___ERRORLOG( QString( "Syntax: grab <widget> [as <filename>] [size w d] [offset x y ]." ) );
-			} else {
-				words.pop_front();
-				QString sWidgetName = words[0];
-				words.pop_front();
-				QRect rect( 0, 0, -1, -1 );
-				QString sFileName = QString( "%1.png" ).arg( sWidgetName );
-				while ( !words.empty() ) {
-					if ( words[0] == "as" ) {
-						words.pop_front();
-						if ( words.size() < 1 ) {
-							___ERRORLOG( QString( "Syntax: grab ... as <filename>" ) );
-						} else {
-							sFileName = words[0];
-							words.pop_front();
-						}
-					} else if ( words[0] == "size" ) {
-						words.pop_front();
-						if ( words.size() < 2 ) {
-							___ERRORLOG( QString( "Syntax: grab ... size <width> <height>" ) );
-						} else {
-							rect.setWidth( words[0].toInt() );
-							rect.setHeight( words[1].toInt() );
-							words.pop_front();
-							words.pop_front();
-						}
-					} else if ( words[0] == "offset" ) {
-						words.pop_front();
-						if ( words.size() < 2 ) {
-							___ERRORLOG( QString( "Syntax: grab ... offset <width> <height>" ) );
-						} else {
-							rect.setX( words[0].toInt() );
-							rect.setY( words[1].toInt() );
-							words.pop_front();
-							words.pop_front();
-						}
-					} else {
-						___ERRORLOG( QString( "Syntax: grab <widget> [as <filename>] [size w d] [offset x y ]."
-											  " Unexpected '%1'" ).arg( words[0] ) );
-						break;
-					}
-				}
-
-				QWidget *pWidget = findWidget( sWidgetName );
-				if ( pWidget ) {
-					QPixmap p = pWidget->grab();
-					QRect oldRect = rect;
-					// Scale 'rect' up to match device pixels of pixmap
-					rect = QRect( rect.topLeft() * p.devicePixelRatio(),
-								  rect.size() * p.devicePixelRatio() );
-					if ( rect.width() <= 0 ) {
-						rect.setWidth( p.rect().width() );
-					}
-					if ( rect.height() <= 0 ) {
-						rect.setHeight( p.rect().height() );
-					}
-					QRect grabRect = rect.intersected( p.rect() );
-					p.copy( grabRect ).save( sFileName );
-					___INFOLOG( QString( "Saved grabbed widget %1" ).arg( sFileName ) );
-				} else {
-					___ERRORLOG( QString( "Couldn't find widget named '%1' to grab" ).arg( sWidgetName ) );
-				}
-			}
-
-		} else if ( sCmd.compare( "slot", Qt::CaseInsensitive ) == 0 ) {
-
-			if ( words.size() >= 3 ) {
-				QString sWidgetName = words[ 1 ];
-				QString sMethodName = words[ 2 ];
-				QWidget *pWidget = findWidget( sWidgetName );
-
-				if ( pWidget ) {
-					___INFOLOG( QString( "Invoking '%1' on '%2'" ).arg( sMethodName, sWidgetName ) );
-					bool bSuccess = false;
-					switch ( words.size() ) {
-					case 3:
-						bSuccess = QMetaObject::invokeMethod( pWidget, sMethodName.toLocal8Bit().data(), Qt::DirectConnection );
-						break;
-					case 4:
-						bSuccess = QMetaObject::invokeMethod( pWidget, sMethodName.toLocal8Bit().data(), Qt::DirectConnection,
-															  Arg( words[3] ) );
-						break;
-					case 5:
-						bSuccess = QMetaObject::invokeMethod( pWidget, sMethodName.toLocal8Bit().data(), Qt::DirectConnection,
-															  Arg( words[3] ), Arg( words[4] ) );
-						break;
-					case 6:
-						bSuccess = QMetaObject::invokeMethod( pWidget, sMethodName.toLocal8Bit().data(), Qt::DirectConnection,
-															  Arg( words[3] ), Arg( words[4] ), Arg( words[5] ) );
-						break;
-					default:
-						___ERRORLOG( "Unsupported number of arguments in %0" );
-					}
-
-					if ( !bSuccess ) {
-						___ERRORLOG( QString( "Couldn't invoke '%1' on '%2'" )
-									 .arg( sMethodName, sWidgetName ) );
-					} else {
-						___INFOLOG( "OK" );
-					}
-				} else {
-					___ERRORLOG( QString( "Couldn't find widget named '%1' to invoke '%2' on" )
-								 .arg( sWidgetName, sMethodName ) );
-				}
-			} else {
-				___ERRORLOG( QString( "Syntax: slot <widget> <method> [args]" ) );
-			}
-		} else {
-			___ERRORLOG( QString("Unknown command '%1'").arg( sCmd ) );
-		}
-	}
-  
-public:
-	ShotList( QStringList shots ) {
-		m_shots = shots;
-	}
-	ShotList( QString sShotsFilename ) {
-		QFile shots( sShotsFilename );
-		if ( ! shots.open( QIODevice::ReadOnly ) ) {
-			___ERRORLOG( QString( "Cannot open shot list file '%1' " ).arg( shots.fileName() ) );
-			return;
-		}
-		while (! shots.atEnd() ) {
-			m_shots << shots.readLine();
-		}
-	}
-
-	void shoot() {
-		m_nNextShot = 0;
-		if ( m_shots.size() != 0 ) {
-			nextShot();
-		}
-	}
-
-public slots:
-	void nextShot( void ) {
-		if ( ( m_nNextShot + 1) < m_shots.size() ) {
-			// Cue up next shot
-			QMetaObject::invokeMethod( this, &ShotList::nextShot, Qt::QueuedConnection );
-		}
-		shoot( m_shots[ m_nNextShot++ ] );
-	}
-
-};
 
 
 int main(int argc, char *argv[])
